requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.10
redis==5.2.0
pyjwt==2.9.0
django-environ==0.11.2
gunicorn==23.0.0
drf-yasg==1.21.8
cryptography==43.0.3
coverage==7.6.4
pytest==8.3.3
pytest-django==4.9.0